Removed: Views no longer have a baseHTML property and no longer fire the onGotBaseHTML and onInsertBaseHTML events.
They also no longer have a fetchContent method.

The need for a default wrapping template, ie. the "Default.html" file. The framework now generates this automatically.

insertBaseHTML and getBaseHTML methods have been removed from the view.

Added:
If your view handler needs to work with templates, specify them in the new "templates" property. This is an object. The key is the reference to fetch the template with. The value is the name of the file (/ if it lives within a sub folder in the html folder).

A new private method "fetchTemplateContent" preloads all the templates based on the "templates" property provided.

A new event "onTemplatesLoaded",is fired when all templates have been loaded.

A new method "getTemplate()" is available to fetch the contents of a loaded template. Make sure this is only ever called after the onTemplatesLoaded event has been fired.

New private method "insertBaseContainer" which will generate the base container div.
Note that this is done BEFORE any templates are loaded. This means that by the time the "onTemplatesLoaded" event has been fired, the view is actually in a ready state and can be used.

OF NOTE:
the "onTemplatesLoaded" event is fired BEFORE view ready events. This is so that you have a chance to setup the view before manipulating it with store data. Be mindful to not use any async functionality in this section as it could throw things out of sync and cause the "onReadyState" event to be fired before you expect it to.